requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-multipart
pydantic
python-dotenv
httpx
